+++ app/Http/Controllers/TaskCommentController.php
@@ -5,7 +5,9 @@ namespace App\Http\Controllers;
 use App\Http\Requests\ValidateTaskComment;
 use App\Http\Requests\ValidateTaskCommentId;
 use App\Http\Requests\ValidateTaskCommentParams;
+use App\Models\TaskComment;
 use App\Services\TaskCommentService;
+use Illuminate\Support\Facades\Gate;
 
 class TaskCommentController extends Controller
 {
@@ -18,52 +20,94 @@ class TaskCommentController extends Controller
 
     public function data(ValidateTaskCommentParams $request) {
         $result = $this->task_comment_service->get($request->task_id);
-        echo json_encode($result->toArray());
+        $result = $result->map(function($item,$key) use($request){
+            $task_comment = TaskComment::find($item['id']);
+            $item['can_update'] = $request->user()->can('update',$task_comment);
+            $item['can_delete'] = $request->user()->can('delete',$task_comment);
+            return $item;
+        });
+        return response()->json($result);
     }
 
     public function edit(ValidateTaskCommentId $request) {
-        $result = $this->task_comment_service->detail($request->id);
+        $result = null;
+        if($request->user()->can('update',TaskComment::find($request->id))) {
+            $result = $this->task_comment_service->detail($request->id);
+        } else {
+            return response()->json([
+                'status' => false,
+                'message' => __('response.not_authorized'),
+                'data' => $result
+            ],403);
+        }
+
         if($result) {
             return [
                 'status' => true,
-                'message' => 'Data retrieved successfully',
+                'message' => __('response.retrieve_succeed'),
                 'data' => $result
             ];
         }
         return [
             'status' => false,
-            'message' => 'Failed to retrieve data',
-            'errors' => []
+            'message' => __('response.retrieve_failed'),
         ];
     }
 
     public function save(ValidateTaskComment $request) {
+        if($request->id == null
+        && !($response = Gate::inspect('create',App\Models\TaskComment::class))->allowed()) {
+            return response()->json([
+                'status' => false,
+                'message' => $response->message()
+            ],403);
+        }
+        if($request->id != null
+        && !($response = Gate::inspect('update',TaskComment::find($request->id)))->allowed()) {
+            return response()->json([
+                'status' => false,
+                'message' => $response->message()
+            ],403);
+        }
+
         $attr = $request->validated();
         unset($attr['task_id']);
         $result = $this->task_comment_service->save($request->task_id, $attr);
-        if(isset($result)) return [
-            'status' => true,
-            'message' => 'Data saved successfully',
-            'data' => $result
-        ];
+
+        if(isset($result)) {
+            $result->can_update = $request->user()->can('update',$result);
+            $result->can_delete = $request->user()->can('delete',$result);
+            return [
+                'status' => true,
+                'message' => __('response.save_succeed'),
+                'data' => $result
+            ];
+        }
         return [
             'status' => false,
-            'message' => 'Failed to save data',
+            'message' => __('response.save_failed'),
             'errors' => []
         ];
     }
 
     public function delete(ValidateTaskCommentId $request) {
+        $response = Gate::inspect('delete',TaskComment::find($request->id));
+        if(!$response->allowed()) {
+            return response()->json([
+                'status' => false,
+                'message' => $response->message()
+            ],403);
+        }
         $result = $this->task_comment_service->delete($request->id);
         if($result) {
             return [
                 'status' => true,
-                'message' => 'Data deleted successfully',
+                'message' => __('response.delete_succeed'),
             ];
         }
         return [
             'status' => false,
-            'message' => 'Failed to delete data'
+            'message' => __('response.delete_failed')
         ];
     }
 }
