+++ app/Http/Controllers/TaskFileController.php
@@ -5,7 +5,9 @@ namespace App\Http\Controllers;
 use App\Http\Requests\ValidateTaskFile;
 use App\Http\Requests\ValidateTaskFileId;
 use App\Http\Requests\ValidateTaskFileParams;
+use App\Models\Task;
 use App\Services\TaskFileService;
+use Illuminate\Support\Facades\Gate;
 
 class TaskFileController extends Controller
 {
@@ -18,7 +20,11 @@ class TaskFileController extends Controller
 
     public function data(ValidateTaskFileParams $request) {
         $result = $this->task_file_service->get($request->task_id);
-        echo json_encode($result->toArray());
+        $result = $result->map(function($item,$key) use($request){
+            $item['can_delete'] = $request->user()->can('deleteFile',[Task::find($request->task_id),$item['id']]);
+            return $item;
+        });
+        return response()->json($result);
     }
 
     public function download(ValidateTaskFileId $request) {
@@ -28,11 +34,15 @@ class TaskFileController extends Controller
 
     public function save(ValidateTaskFile $request) {
         $result = $this->task_file_service->save($request->task_id, $request->file('file'));
-        if(isset($result)) return [
-            'status' => true,
-            'message' => 'Data saved successfully',
-            'data' => $result
-        ];
+        if(isset($result)) {
+            $task = Task::find($request->task_id);
+            $result->can_delete = $request->user()->can('deleteFile',[$task, $result->id]);
+            return [
+                'status' => true,
+                'message' => 'Data saved successfully',
+                'data' => $result
+            ];
+        }
         return [
             'status' => false,
             'message' => 'Failed to save data',
@@ -41,6 +51,11 @@ class TaskFileController extends Controller
     }
 
     public function delete(ValidateTaskFileId $request) {
+        $task = $this->task_file_service->getTask($request->id);
+        $response = Gate::inspect('deleteFile',[$task, $request->id]);
+        if(!$response->allowed())
+            return response()->json(['status' => false, 'message' => $response->message()], 403);
+
         $result = $this->task_file_service->delete($request->id);
         if($result) {
             return [
